re-structure folders in product to better structure components and to be more organized
