+++ core/components/product/index.ts
@@ -1,5 +1,5 @@
-export { ProductDetailBike } from './product-detail-bike';
-export { ProductDetailScooter } from './product-detail-scooter';
-export { getProductDetailVariant } from './product-detail-router';
-export type { ProductDetailBikeProps } from './product-detail-bike';
-export type { ProductDetailScooterProps } from './product-detail-scooter';
+export { ProductDetailBike } from './bike/product-detail-bike';
+export { ProductDetailScooter } from './scooter/product-detail-scooter';
+export { getProductDetailVariant } from './layout/product-detail-router';
+export type { ProductDetailBikeProps } from './bike/product-detail-bike';
+export type { ProductDetailScooterProps } from './scooter/product-detail-scooter';
