+++ core/components/product/layout/product-detail-layout.tsx
