+++ core/components/product/scooter/product-detail-scooter.tsx
